Java 1-3 01~04 first commit

+++ src/main/java/method/q06/Circle.java
@@ -3,13 +3,13 @@ package method.q06;
 public class Circle {
 
 	public static void main(String[] args) {
-		int num = 5;
+		double num = 5.0;
 		double area = getCircleArea(num);
 		System.out.println("半径：" + num + ".0");
 		System.out.println("円の面積：" + area);
 	}
 
-	public static double getCircleArea(int num) {
+	public static double getCircleArea(double num) {
 		return (num * num) * 3.14;
 	}
 }
